+++ migrations/000001_create_table_profile.up.sql
@@ -1,8 +1,7 @@
 create table profile (id uuid primary key, 
     login varchar not null, 
     password varchar not null, 
-    balance int not null,
-    withdrawn int not null,
+    balance numeric(10, 2) not null,
+    withdrawn numeric(10, 2) not null,
     CONSTRAINT login_unique UNIQUE(login));
 
-
